Set the margin-top of the separator instead of messing with flex elements
Use a different approach to fix issue #1. Instead of trying to work around the flex elements, and create a proper hierarchy, the simplest fix for the issue would be to set the `margin-top` of the separator, instead of setting the `height` of one of the two flex elements above the separator.

This way the two elements above are not changed, and the browser calculates its location correctly, keeping them aligned and solving the issue.